requirements.txt
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.5.0
numpy>=1.19.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
pillow>=8.0.0
pandas>=1.2.0
tqdm>=4.50.0
glob2>=0.7